+++ NEWS.md
@@ -7,6 +7,7 @@ EDCimport is a package designed to easily import data from EDC software TrialMas
 ### New features
 
 -   New arguments in `edc_swimmerplot()`: `origin_fun` to summarise `origin` at patient level using, and `data_list` to control the datasets.
+-   New function `compare_databases()`, which compares the structure of several extractions of a database: added/removed columns, number of patients, etc (#26). See the examples for a demo.
 -   New features in `edc_viewer()`:
     -   Support for multiple instances on different ports with custom datasets (#100, #114)\
         For instance, you can now run `edc_viewer(data=lst(iris, mtcars), port=1212)`
@@ -14,6 +15,8 @@ EDCimport is a package designed to easily import data from EDC software TrialMas
 
 ### Bug fixes & Improvements
 
+-   `edc_viewer()`: support for multiple instances on different ports(#114).
+-   `edc_viewer()`: new button to browse all the column labels (#113).
 -   Fixed modifiers `edc_clean_names()`, `edc_unify_subjid()`, and `edc_split_mixed()` that stripped attributes like project name (#111).
 -   `edc_data_stop()` now works without a SUBJID and defaults to no issue number (#109).
 -   Fixed bugs in `edc_left_join()` with case-sensitivity on SUBJID (#108, #117).
@@ -24,6 +27,7 @@ EDCimport is a package designed to easily import data from EDC software TrialMas
 -   Improved `edc_warn_extraction_date()` with a strict unit "days".
 -   Improved `save_plotly()` with a glue syntax for param `file`.
 
+
 # EDCimport 0.6.0
 
 ### Documentation
@@ -32,7 +36,7 @@ EDCimport is a package designed to easily import data from EDC software TrialMas
 
 ### New features
 
--   New functions `edc_patient_gridplot()`, which creates a ggplot matrix giving the presence of all patients in all datasets (#77)
+-   New function `edc_patient_gridplot()`, which creates a ggplot matrix giving the presence of all patients in all datasets (#77)
 -   New functions `edc_left_join()`, `edc_right_join()`, and `edc_full_join()`, which perform joins with defaults to subject ID as primary key (#82)
 -   New function `edc_viewer()`, which runs a shiny application for easily browsing your database (#83)
 -   New function `set_project_name()`, to set the project name when reading from a directory (#96)